requirements_test.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
billiard==4.2.4
cachetools==5.5.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.123.5
frozenlist==1.8.0
google-ai-generativelanguage==0.1.0
google-api-core==2.28.1
google-auth==2.43.0
google-genai==1.53.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.72.0
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn>=21.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.9.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==1.1.2
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.1
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.4.56
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.5
openai==2.2.0
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
passlib==1.7.4
pgvector==0.4.2
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.27.2
pypdf==6.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
PyYAML==6.0.3
redis==6.1.1
requests==2.32.5
requests-toolbelt==1.0.0
rq==2.3.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.44.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
uuid_utils==0.12.0
uvicorn==0.33.0
uvloop==0.22.1
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.14
websockets==13.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0